+++ src/main/java/webserver/http/HttpVersion.java
@@ -1,4 +1,4 @@
-package webserver.http.request;
+package webserver.http;
 
 import java.util.HashMap;
 import java.util.Map;
