+++ src/main/java/webserver/http/Pair.java
@@ -1,4 +1,4 @@
-package webserver.http.request;
+package webserver.http;
 
 import java.util.Objects;
